+++ typescript/examples/vercel-ai/viem/index.ts
@@ -1,14 +1,17 @@
+import readline from "node:readline";
+
 import { openai } from "@ai-sdk/openai";
 import { generateText } from "ai";
 
 import { http } from "viem";
 import { createWalletClient } from "viem";
 import { privateKeyToAccount } from "viem/accounts";
-import { sepolia } from "viem/chains";
+import { base } from "viem/chains";
 
 import { getOnChainTools } from "@goat-sdk/adapter-vercel-ai";
 import { PEPE, USDC, erc20 } from "@goat-sdk/plugin-erc20";
 
+import { uniswap } from "@goat-sdk/plugin-uniswap";
 import { sendETH } from "@goat-sdk/wallet-evm";
 import { viem } from "@goat-sdk/wallet-viem";
 
@@ -19,21 +22,58 @@ const account = privateKeyToAccount(process.env.WALLET_PRIVATE_KEY as `0x${strin
 const walletClient = createWalletClient({
     account: account,
     transport: http(process.env.RPC_PROVIDER_URL),
-    chain: sepolia,
+    chain: base,
 });
 
 (async () => {
     const tools = await getOnChainTools({
         wallet: viem(walletClient),
-        plugins: [sendETH(), erc20({ tokens: [USDC, PEPE] })],
+        plugins: [
+            sendETH(),
+            erc20({ tokens: [USDC, PEPE] }),
+            uniswap({
+                baseUrl: process.env.UNISWAP_BASE_URL as string,
+                apiKey: process.env.UNISWAP_API_KEY as string,
+            }),
+        ],
     });
 
-    const result = await generateText({
-        model: openai("gpt-4o-mini"),
-        tools: tools,
-        maxSteps: 5,
-        prompt: "Get the balance of the USDC token",
+    const rl = readline.createInterface({
+        input: process.stdin,
+        output: process.stdout,
     });
 
-    console.log(result.text);
+    while (true) {
+        const prompt = await new Promise<string>((resolve) => {
+            rl.question('Enter your prompt (or "exit" to quit): ', resolve);
+        });
+
+        if (prompt === "exit") {
+            rl.close();
+            break;
+        }
+
+        console.log("\n-------------------\n");
+        console.log("TOOLS CALLED");
+        console.log("\n-------------------\n");
+
+        console.log("\n-------------------\n");
+        console.log("RESPONSE");
+        console.log("\n-------------------\n");
+        try {
+            const result = await generateText({
+                model: openai("gpt-4o-mini"),
+                tools: tools,
+                maxSteps: 10,
+                prompt: prompt,
+                onStepFinish: (event) => {
+                    console.log(event.toolResults);
+                },
+            });
+            console.log(result.text);
+        } catch (error) {
+            console.error(error);
+        }
+        console.log("\n-------------------\n");
+    }
 })();