+++ typescript/packages/core/src/types/Chain.ts
@@ -2,7 +2,19 @@
  * @param type - "evm" or "solana", extend this union as needed (e.g., "sui")
  * @param id - Chain ID, optional for EVM
  */
-export type Chain = EvmChain | SolanaChain | AptosChain | ChromiaChain | FuelChain;
+export type Chain =
+    | EvmChain
+    | SolanaChain
+    | AptosChain
+    | ChromiaChain
+    | FuelChain
+    | SuiChain
+    | ZilliqaChain
+    | CosmosChain;
+
+export type SuiChain = {
+    type: "sui";
+};
 
 export type EvmChain = {
     type: "evm";
@@ -23,4 +35,14 @@ export type ChromiaChain = {
 
 export type FuelChain = {
     type: "fuel";
+};
+
+export type ZilliqaChain = {
+    type: "zilliqa";
+    id: number;
+    evmId: number;
+};
+
+export type CosmosChain = {
+    type: "cosmos";
 };