+++ typescript/packages/plugins/ionic/src/ionic.plugin.ts
@@ -1,17 +1,11 @@
 import { type Chain, PluginBase } from "@goat-sdk/core";
 import { EVMWalletClient } from "@goat-sdk/wallet-evm";
-import { mode } from "viem/chains";
-import { base } from "viem/chains";
-import { optimism } from "viem/chains";
+import { base, mode, optimism } from "viem/chains";
 import { IonicService } from "./ionic.service";
 
-interface IonicPluginOptions {
-    supportedTokens?: string[]; // Optional list of tokens to support
-}
-
 export class IonicPlugin extends PluginBase<EVMWalletClient> {
-    constructor({ supportedTokens }: IonicPluginOptions = {}) {
-        super("ionic", [new IonicService(supportedTokens)]);
+    constructor() {
+        super("ionic", [new IonicService()]);
     }
 
     supportsChain(chain: Chain): boolean {
@@ -19,6 +13,6 @@ export class IonicPlugin extends PluginBase<EVMWalletClient> {
     }
 }
 
-export function ionic(options: IonicPluginOptions = {}) {
-    return new IonicPlugin(options);
+export function ionic() {
+    return new IonicPlugin();
 }