+++ typescript/examples/langchain/viem/CHANGELOG.md
@@ -1,5 +1,52 @@
 # goat-examples-langchain-viem
 
+## 0.2.11
+
+### Patch Changes
+
+- Updated dependencies [2b4b8e8]
+  - @goat-sdk/core@0.4.6
+  - @goat-sdk/adapter-langchain@0.2.9
+  - @goat-sdk/plugin-erc20@0.2.8
+  - @goat-sdk/wallet-evm@0.2.6
+  - @goat-sdk/wallet-viem@0.2.6
+
+## 0.2.10
+
+### Patch Changes
+
+- Updated dependencies [b9af25b]
+  - @goat-sdk/core@0.4.5
+  - @goat-sdk/adapter-langchain@0.2.8
+  - @goat-sdk/plugin-erc20@0.2.7
+  - @goat-sdk/wallet-evm@0.2.5
+  - @goat-sdk/wallet-viem@0.2.5
+
+## 0.2.9
+
+### Patch Changes
+
+- Updated dependencies [b9317ec]
+  - @goat-sdk/adapter-langchain@0.2.7
+
+## 0.2.8
+
+### Patch Changes
+
+- Updated dependencies [50180d4]
+  - @goat-sdk/core@0.4.4
+  - @goat-sdk/adapter-langchain@0.2.6
+  - @goat-sdk/plugin-erc20@0.2.6
+  - @goat-sdk/wallet-evm@0.2.4
+  - @goat-sdk/wallet-viem@0.2.4
+
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies [1906022]
+  - @goat-sdk/adapter-langchain@0.2.5
+
 ## 0.2.6
 
 ### Patch Changes