+++ typescript/examples/vercel-ai/fuel/CHANGELOG.md
@@ -1,5 +1,39 @@
 # goat-examples-vercel-ai-fuel
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [2b4b8e8]
+  - @goat-sdk/core@0.4.6
+  - @goat-sdk/adapter-vercel-ai@0.2.7
+  - @goat-sdk/wallet-fuel@0.1.6
+
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [b9af25b]
+  - @goat-sdk/core@0.4.5
+  - @goat-sdk/adapter-vercel-ai@0.2.6
+  - @goat-sdk/wallet-fuel@0.1.5
+
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies [50180d4]
+  - @goat-sdk/core@0.4.4
+  - @goat-sdk/adapter-vercel-ai@0.2.5
+  - @goat-sdk/wallet-fuel@0.1.4
+
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [1906022]
+  - @goat-sdk/adapter-vercel-ai@0.2.4
+
 ## 0.1.3
 
 ### Patch Changes