+++ typescript/examples/vercel-ai/solana/CHANGELOG.md
@@ -1,5 +1,66 @@
 # goat-examples-vercel-ai-solana
 
+## 0.2.15
+
+### Patch Changes
+
+- Updated dependencies [d8b05a5]
+  - @goat-sdk/plugin-coingecko@0.2.7
+
+## 0.2.14
+
+### Patch Changes
+
+- Updated dependencies [2b4b8e8]
+  - @goat-sdk/core@0.4.6
+  - @goat-sdk/adapter-vercel-ai@0.2.7
+  - @goat-sdk/plugin-coingecko@0.2.6
+  - @goat-sdk/plugin-jupiter@0.2.11
+  - @goat-sdk/plugin-pumpfun@0.1.4
+  - @goat-sdk/plugin-spl-token@0.2.13
+  - @goat-sdk/wallet-solana@0.2.11
+
+## 0.2.13
+
+### Patch Changes
+
+- Updated dependencies [b9af25b]
+  - @goat-sdk/core@0.4.5
+  - @goat-sdk/adapter-vercel-ai@0.2.6
+  - @goat-sdk/plugin-coingecko@0.2.5
+  - @goat-sdk/plugin-jupiter@0.2.10
+  - @goat-sdk/plugin-pumpfun@0.1.3
+  - @goat-sdk/plugin-spl-token@0.2.12
+  - @goat-sdk/wallet-solana@0.2.10
+
+## 0.2.12
+
+### Patch Changes
+
+- Updated dependencies [f159a7e]
+- Updated dependencies [50180d4]
+  - @goat-sdk/plugin-coingecko@0.2.4
+  - @goat-sdk/core@0.4.4
+  - @goat-sdk/adapter-vercel-ai@0.2.5
+  - @goat-sdk/plugin-jupiter@0.2.9
+  - @goat-sdk/plugin-pumpfun@0.1.2
+  - @goat-sdk/plugin-spl-token@0.2.11
+  - @goat-sdk/wallet-solana@0.2.9
+
+## 0.2.11
+
+### Patch Changes
+
+- Updated dependencies [1906022]
+  - @goat-sdk/adapter-vercel-ai@0.2.4
+
+## 0.2.10
+
+### Patch Changes
+
+- Updated dependencies [7f3c8dc]
+  - @goat-sdk/plugin-spl-token@0.2.10
+
 ## 0.2.9
 
 ### Patch Changes