+++ typescript/examples/vercel-ai/viem/CHANGELOG.md
@@ -1,5 +1,70 @@
 # goat-examples-vercel-ai-viem
 
+## 0.2.12
+
+### Patch Changes
+
+- 5b08d6c: Release 0x plugin
+
+## 0.2.11
+
+### Patch Changes
+
+- d8b05a5: Fix endpoints
+- Updated dependencies [d8b05a5]
+  - @goat-sdk/plugin-uniswap@0.2.9
+
+## 0.2.10
+
+### Patch Changes
+
+- bff4fa1: Minor chores
+- Updated dependencies [bff4fa1]
+  - @goat-sdk/plugin-uniswap@0.2.8
+
+## 0.2.9
+
+### Patch Changes
+
+- Updated dependencies [2b4b8e8]
+  - @goat-sdk/core@0.4.6
+  - @goat-sdk/adapter-vercel-ai@0.2.7
+  - @goat-sdk/plugin-erc20@0.2.8
+  - @goat-sdk/plugin-erc721@0.1.17
+  - @goat-sdk/wallet-evm@0.2.6
+  - @goat-sdk/wallet-viem@0.2.6
+
+## 0.2.8
+
+### Patch Changes
+
+- Updated dependencies [b9af25b]
+  - @goat-sdk/core@0.4.5
+  - @goat-sdk/adapter-vercel-ai@0.2.6
+  - @goat-sdk/plugin-erc20@0.2.7
+  - @goat-sdk/plugin-erc721@0.1.16
+  - @goat-sdk/wallet-evm@0.2.5
+  - @goat-sdk/wallet-viem@0.2.5
+
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies [50180d4]
+  - @goat-sdk/core@0.4.4
+  - @goat-sdk/adapter-vercel-ai@0.2.5
+  - @goat-sdk/plugin-erc20@0.2.6
+  - @goat-sdk/plugin-erc721@0.1.15
+  - @goat-sdk/wallet-evm@0.2.4
+  - @goat-sdk/wallet-viem@0.2.4
+
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies [1906022]
+  - @goat-sdk/adapter-vercel-ai@0.2.4
+
 ## 0.2.5
 
 ### Patch Changes