+++ typescript/packages/plugins/coingecko/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @goat-sdk/plugin-coingecko
 
+## 0.2.7
+
+### Patch Changes
+
+- d8b05a5: Fix endpoints
+
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies [2b4b8e8]
+  - @goat-sdk/core@0.4.6
+
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies [b9af25b]
+  - @goat-sdk/core@0.4.5
+
+## 0.2.4
+
+### Patch Changes
+
+- f159a7e: feat: add new Coingecko endpoints (search, coin price by contract address, coin data, supported coins, historical data, OHLC) and unify request code. Includes parameter improvements and API key header support.
+- Updated dependencies [50180d4]
+  - @goat-sdk/core@0.4.4
+
 ## 0.2.3
 
 ### Patch Changes