+++ typescript/packages/plugins/erc721/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @goat-sdk/plugin-erc721
 
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies [2b4b8e8]
+  - @goat-sdk/core@0.4.6
+  - @goat-sdk/wallet-evm@0.2.6
+
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies [b9af25b]
+  - @goat-sdk/core@0.4.5
+  - @goat-sdk/wallet-evm@0.2.5
+
+## 0.1.15
+
+### Patch Changes
+
+- Updated dependencies [50180d4]
+  - @goat-sdk/core@0.4.4
+  - @goat-sdk/wallet-evm@0.2.4
+
 ## 0.1.14
 
 ### Patch Changes