+++ typescript/packages/plugins/spl-token/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @goat-sdk/plugin-spl-token
 
+## 0.2.13
+
+### Patch Changes
+
+- Updated dependencies [2b4b8e8]
+  - @goat-sdk/core@0.4.6
+  - @goat-sdk/wallet-solana@0.2.11
+
+## 0.2.12
+
+### Patch Changes
+
+- Updated dependencies [b9af25b]
+  - @goat-sdk/core@0.4.5
+  - @goat-sdk/wallet-solana@0.2.10
+
+## 0.2.11
+
+### Patch Changes
+
+- Updated dependencies [50180d4]
+  - @goat-sdk/core@0.4.4
+  - @goat-sdk/wallet-solana@0.2.9
+
+## 0.2.10
+
+### Patch Changes
+
+- 7f3c8dc: Fix base units error
+
 ## 0.2.9
 
 ### Patch Changes