+++ typescript/packages/wallets/crossmint/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @goat-sdk/crossmint
 
+## 0.3.13
+
+### Patch Changes
+
+- Updated dependencies [2b4b8e8]
+  - @goat-sdk/core@0.4.6
+  - @goat-sdk/wallet-evm@0.2.6
+  - @goat-sdk/wallet-solana@0.2.11
+
+## 0.3.12
+
+### Patch Changes
+
+- Updated dependencies [b9af25b]
+  - @goat-sdk/core@0.4.5
+  - @goat-sdk/wallet-evm@0.2.5
+  - @goat-sdk/wallet-solana@0.2.10
+
+## 0.3.11
+
+### Patch Changes
+
+- Updated dependencies [50180d4]
+  - @goat-sdk/core@0.4.4
+  - @goat-sdk/wallet-evm@0.2.4
+  - @goat-sdk/wallet-solana@0.2.9
+
 ## 0.3.10
 
 ### Patch Changes