+++ typescript/packages/wallets/viem/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @goat-sdk/wallet-viem
 
+## 0.2.6
+
+### Patch Changes
+
+- @goat-sdk/wallet-evm@0.2.6
+
+## 0.2.5
+
+### Patch Changes
+
+- @goat-sdk/wallet-evm@0.2.5
+
+## 0.2.4
+
+### Patch Changes
+
+- @goat-sdk/wallet-evm@0.2.4
+
 ## 0.2.3
 
 ### Patch Changes